backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
langchain==0.3.26          # -> core >= 0.3.68
langchain-openai==0.3.28   # intégrations OpenAI
langgraph==0.5.3
duckduckgo-search==8.1.1
tenacity==8.3.0
ddgs==0.1.0